Create 4.25 LAB: Palindrome
A palindrome is a word or a phrase that is the same when read both forward and backward. Examples are: "bob," "sees," or "never odd or even" (ignoring spaces). Write a program whose input is a word or phrase, and that outputs whether the input is a palindrome.

Ex: If the input is:
bob

the output is:
palindrome: bob

Ex: If the input is:
bobby

the output is:
not a palindrome: bobby