Create 4.40 PRACTICE: Loops (nested): HTML table
HTML is the language of web pages. Items start and end with tags. A table starts with <table> and ends with </table>. In a table, a row starts with <tr> and ends with </tr>. In a row, a column starts with <td> and ends with </td>. Given two integers for rows and columns, generate an appropriately sized html table. Place the character c in each table cell. If the input is 2 3, the output is:

<table>
<tr> <td> c </td> <td> c </td> <td> c </td> </tr>
<tr> <td> c </td> <td> c </td> <td> c </td> </tr>
</table>