Create 5.20 LAB: Middle item
Given a sorted list of integers, output the middle integer. A negative number indicates the end of the input (the negative number is not a part of the sorted list). Assume the number of integers is always odd.

Ex: If the input is:
2 3 4 8 11 -1 

the output is:
Middle item: 4

The maximum number of inputs for any test case should not exceed 9. If exceeded, output "Too many numbers".