Create 5.23 LAB: Word frequencies
Write a program that reads a list of words. Then, the program outputs those words and their frequencies. The input begins with an integer indicating the number of words that follow. Assume that the list will always contain fewer than 20 words.

Ex: If the input is:
5 hey hi Mark hi mark

the output is:
hey - 1
hi - 2
Mark - 1
hi - 2
mark - 1