Create 5.26 LAB: Two smallest numbers
Write a program that reads a list of integers, and outputs the two smallest integers in the list, in ascending order. The input begins with an integer indicating the number of integers that follow.

Ex: If the input is:
5 10 5 3 21 2

the output is:
2 and 3

You can assume that the list of integers will have at least 2 values.

To achieve the above, first read the integers into a vector.