Create 5.39 PRACTICE: Vectors: Reverse in place
Reversing a vector is a common task. One approach copies to a second vector in reverse, then copies the second vector back to the first. However, to save space, reversing a vector without using a second vector is sometimes preferable. Write a function that reverses a given vector, without using a second vector. If the original vector's values are 2 5 9 7, the vector after reversing is 7 9 5 2.