Create 6.39 LAB: Word frequencies - functions
Write a program that reads a list of words. Then, the program outputs those words and their frequencies. The input begins with an integer indicating the number of words that follow.

Ex: If the input is:
5 hey hi Mark hi mark

the output is:
hey 1
hi 2
Mark 1
hi 2
mark 1


Your program must define and use the following function:
int GetFrequencyOfWord(vector<string> wordsList, string currWord)