Create 6.44 LAB: Fibonacci sequence
The Fibonacci sequence begins with 0 and then 1 follows. All subsequent values are the sum of the previous two, for example: 0, 1, 1, 2, 3, 5, 8, 13. Complete the Fibonacci() function, which has an index, n, as parameter and returns the nth value in the sequence. Any negative index values should return -1.

Ex: If the input is:
7

the output is:
Fibonacci(7) is 13

Note: Use a for loop and DO NOT use recursion.